[modified] variable in DataTransform and DTWithParent for instances that the value limit is exceeded

+++ MMOCoreORB/src/server/zone/packets/object/DataTransform.h
@@ -74,7 +74,7 @@ class DataTransformCallback : public MessageCallback {
 	ValidatedPosition validPosition;
 	Transform transform;
 
-	int deltaTime;
+	long deltaTime;
 
 public:
 	DataTransformCallback(ObjectControllerMessageCallback* objectControllerCallback) : MessageCallback(objectControllerCallback->getClient(), objectControllerCallback->getServer()) {
